Refactor the flatrelatedfield, cachefield and jsonfield logic into magic mixins ...
